# openapi.yaml
openapi: 3.0.0
info:
  title: Jira Analytics API
  description: Node.js server for analyzing Jira issues. Updated for Jira Cloud JQL Search API (v3).
  version: 1.1.0

servers:
  - url: http://localhost:3000
    description: Local Development Server

components:
  parameters:
    JiraUrl:
      in: header
      name: x-jira-url
      schema:
        type: string
      required: true
    JiraEmail:
      in: header
      name: x-jira-email
      schema:
        type: string
        format: email
      required: true
    JiraApiKey:
      in: header
      name: x-jira-api-key
      schema:
        type: string
      required: true

paths:
  /api/projects:
    get:
      operationId: listProjects
      summary: List all Jira projects
      parameters:
        - $ref: '#/components/parameters/JiraUrl'
        - $ref: '#/components/parameters/JiraEmail'
        - $ref: '#/components/parameters/JiraApiKey'
      responses:
        '200':
          description: Success

  /api/data-understanding:
    get:
      operationId: getDataUnderstanding
      summary: Get issue statistics
      parameters:
        - $ref: '#/components/parameters/JiraUrl'
        - $ref: '#/components/parameters/JiraEmail'
        - $ref: '#/components/parameters/JiraApiKey'
        - name: projectKey
          in: query
          schema:
            type: string
          required: true
        - name: nextPageToken
          in: query
          description: Token for the next page of results (from previous response).
          schema:
            type: string
        - name: maxResults
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Returns stats and a nextPageToken if more data exists.

  /api/time-analysis:
    get:
      operationId: getTimeAnalysis
      summary: Analyze issue aging and spikes
      parameters:
        - $ref: '#/components/parameters/JiraUrl'
        - $ref: '#/components/parameters/JiraEmail'
        - $ref: '#/components/parameters/JiraApiKey'
        - name: projectKey
          in: query
          required: true
          schema: 
            type: string
        - name: nextPageToken
          in: query
          schema: 
            type: string
        - name: maxResults
          in: query
          schema: 
            type: integer
            default: 50
        - name: x-stalled-days
          in: header
          schema:
            type: integer
            default: 14
      responses:
        '200':
          description: Time-based statistics returned.

  /api/workflow-analysis:
    get:
      operationId: getWorkflowAnalysis
      summary: Analyze transitions and bottlenecks
      parameters:
        - $ref: '#/components/parameters/JiraUrl'
        - $ref: '#/components/parameters/JiraEmail'
        - $ref: '#/components/parameters/JiraApiKey'
        - name: projectKey
          in: query
          required: true
          schema: 
            type: string
        - name: nextPageToken
          in: query
          schema: 
            type: string
        - name: maxResults
          in: query
          schema: 
            type: integer
            default: 50
      responses:
        '200':
          description: Workflow analysis returned.